+++ packages/ai/src/generate-text/content-part.ts
@@ -1,6 +1,7 @@
 import { ProviderMetadata } from '../types';
 import { Source } from '../types/language-model';
 import { GeneratedFile } from './generated-file';
+import { ReasoningOutput } from './reasoning-output';
 import { TypedToolCall } from './tool-call';
 import { TypedToolError } from './tool-error';
 import { TypedToolResult } from './tool-result';
@@ -8,7 +9,7 @@ import { ToolSet } from './tool-set';
 
 export type ContentPart<TOOLS extends ToolSet> =
   | { type: 'text'; text: string; providerMetadata?: ProviderMetadata }
-  | { type: 'reasoning'; text: string; providerMetadata?: ProviderMetadata }
+  | ReasoningOutput
   | ({ type: 'source' } & Source)
   | { type: 'file'; file: GeneratedFile; providerMetadata?: ProviderMetadata } // different because of GeneratedFile object
   | ({ type: 'tool-call' } & TypedToolCall<TOOLS> & {