+++ packages/anthropic/src/anthropic-messages-language-model.test.ts
@@ -3,9 +3,9 @@ import {
   LanguageModelV2StreamPart,
   JSONValue,
 } from '@ai-sdk/provider';
+import { createTestServer } from '@ai-sdk/test-server/with-vitest';
 import {
   convertReadableStreamToArray,
-  createTestServer,
   mockId,
 } from '@ai-sdk/provider-utils/test';
 import { AnthropicProviderOptions } from './anthropic-messages-options';
@@ -553,7 +553,7 @@ describe('AnthropicMessagesLanguageModel', () => {
             providerOptions: {
               anthropic: {
                 cacheControl: { type: 'ephemeral' },
-              },
+              } satisfies AnthropicProviderOptions,
             },
           },
         ],
@@ -679,9 +679,6 @@ describe('AnthropicMessagesLanguageModel', () => {
       const model = provider('claude-3-haiku-20240307');
 
       const result = await model.doGenerate({
-        headers: {
-          'anthropic-beta': 'extended-cache-ttl-2025-04-11',
-        },
         prompt: [
           {
             role: 'user',
@@ -689,7 +686,7 @@ describe('AnthropicMessagesLanguageModel', () => {
             providerOptions: {
               anthropic: {
                 cacheControl: { type: 'ephemeral', ttl: '1h' },
-              },
+              } satisfies AnthropicProviderOptions,
             },
           },
         ],
@@ -2534,9 +2531,6 @@ describe('AnthropicMessagesLanguageModel', () => {
 
       const { stream } = await model.doStream({
         prompt: TEST_PROMPT,
-        headers: {
-          'anthropic-beta': 'extended-cache-ttl-2025-04-11',
-        },
       });
 
       expect(await convertReadableStreamToArray(stream)).toMatchInlineSnapshot(`