+++ packages/openai/src/responses/convert-to-openai-responses-input.ts
@@ -1,15 +1,15 @@
 import {
   LanguageModelV2CallWarning,
   LanguageModelV2Prompt,
+  LanguageModelV2ToolCallPart,
   UnsupportedFunctionalityError,
 } from '@ai-sdk/provider';
-import { parseProviderOptions } from '@ai-sdk/provider-utils';
+import { convertToBase64, parseProviderOptions } from '@ai-sdk/provider-utils';
 import { z } from 'zod/v4';
 import {
-  OpenAIResponsesPrompt,
+  OpenAIResponsesInput,
   OpenAIResponsesReasoning,
 } from './openai-responses-api-types';
-import { convertToBase64 } from '@ai-sdk/provider-utils';
 
 /**
  * Check if a string is a file ID based on the given prefixes
@@ -20,19 +20,21 @@ function isFileId(data: string, prefixes?: readonly string[]): boolean {
   return prefixes.some(prefix => data.startsWith(prefix));
 }
 
-export async function convertToOpenAIResponsesMessages({
+export async function convertToOpenAIResponsesInput({
   prompt,
   systemMessageMode,
   fileIdPrefixes,
+  store,
 }: {
   prompt: LanguageModelV2Prompt;
   systemMessageMode: 'system' | 'developer' | 'remove';
   fileIdPrefixes?: readonly string[];
+  store: boolean;
 }): Promise<{
-  messages: OpenAIResponsesPrompt;
+  input: OpenAIResponsesInput;
   warnings: Array<LanguageModelV2CallWarning>;
 }> {
-  const messages: OpenAIResponsesPrompt = [];
+  const input: OpenAIResponsesInput = [];
   const warnings: Array<LanguageModelV2CallWarning> = [];
 
   for (const { role, content } of prompt) {
@@ -40,11 +42,11 @@ export async function convertToOpenAIResponsesMessages({
       case 'system': {
         switch (systemMessageMode) {
           case 'system': {
-            messages.push({ role: 'system', content });
+            input.push({ role: 'system', content });
             break;
           }
           case 'developer': {
-            messages.push({ role: 'developer', content });
+            input.push({ role: 'developer', content });
             break;
           }
           case 'remove': {
@@ -65,7 +67,7 @@ export async function convertToOpenAIResponsesMessages({
       }
 
       case 'user': {
-        messages.push({
+        input.push({
           role: 'user',
           content: content.map((part, index) => {
             switch (part.type) {
@@ -123,11 +125,12 @@ export async function convertToOpenAIResponsesMessages({
 
       case 'assistant': {
         const reasoningMessages: Record<string, OpenAIResponsesReasoning> = {};
+        const toolCallParts: Record<string, LanguageModelV2ToolCallPart> = {};
 
         for (const part of content) {
           switch (part.type) {
             case 'text': {
-              messages.push({
+              input.push({
                 role: 'assistant',
                 content: [{ type: 'output_text', text: part.text }],
                 id:
@@ -136,11 +139,13 @@ export async function convertToOpenAIResponsesMessages({
               break;
             }
             case 'tool-call': {
+              toolCallParts[part.toolCallId] = part;
+
               if (part.providerExecuted) {
                 break;
               }
 
-              messages.push({
+              input.push({
                 type: 'function_call',
                 call_id: part.toolCallId,
                 name: part.toolName,
@@ -151,11 +156,18 @@ export async function convertToOpenAIResponsesMessages({
               break;
             }
 
+            // assistant tool result parts are from provider-executed tools:
             case 'tool-result': {
-              warnings.push({
-                type: 'other',
-                message: `tool result parts in assistant messages are not supported for OpenAI responses`,
-              });
+              if (store) {
+                // use item references to refer to tool results from built-in tools
+                input.push({ type: 'item_reference', id: part.toolCallId });
+              } else {
+                warnings.push({
+                  type: 'other',
+                  message: `Results for OpenAI tool ${part.toolName} are not sent to the API when store is false`,
+                });
+              }
+
               break;
             }
 
@@ -193,7 +205,7 @@ export async function convertToOpenAIResponsesMessages({
                       providerOptions?.reasoningEncryptedContent,
                     summary: summaryParts,
                   };
-                  messages.push(reasoningMessages[reasoningId]);
+                  input.push(reasoningMessages[reasoningId]);
                 } else {
                   existingReasoningMessage.summary.push(...summaryParts);
                 }
@@ -228,7 +240,7 @@ export async function convertToOpenAIResponsesMessages({
               break;
           }
 
-          messages.push({
+          input.push({
             type: 'function_call_output',
             call_id: part.toolCallId,
             output: contentValue,
@@ -245,7 +257,7 @@ export async function convertToOpenAIResponsesMessages({
     }
   }
 
-  return { messages, warnings };
+  return { input, warnings };
 }
 
 const openaiResponsesReasoningProviderOptionsSchema = z.object({